refactor: adds tabs to improve screens navigation
