+++ src/App.tsx
@@ -31,9 +31,8 @@ import { Login } from "./Screens/Login";
 //Telas privadas
 import { Home } from "./Screens/PrivateScreens/Home";
 import { Noticias } from "./Screens/PrivateScreens/Noticias";
-import { Storage } from "./Screens/PrivateScreens/Estoque/Storage";
-import { Outlet } from "./Screens/PrivateScreens/Estoque/Outlet";
 import { Perfil } from "./Screens/PrivateScreens/Perfil";
+import { StorageHome } from "./Screens/PrivateScreens/Estoque/StorageHome";
 
 function App() {
   return (
@@ -58,8 +57,7 @@ function App() {
             <Route index element={<Home />} />
             <Route path="noticias" element={<Noticias />} />
             <Route path="perfil" element={<Perfil />} />
-            <Route path="estoque" element={<Storage />}/>
-            <Route path="outlet" element={<Outlet />}/>
+            <Route path="estoque" element={<StorageHome />}/>
           </Route>
         </Route>
       </Routes>
