+++ src/components/Storage/index.scss
@@ -10,24 +10,32 @@
     gap: 1rem;
 }
 
+.page-title {
+    font-size: 1.5rem;
+    color: $dark;
+    font-weight: bold;
+    margin-bottom: 1rem;
+}
+
 .storage-header {
     display: flex;
-    justify-content: space-between;
+    justify-content: flex-end;
     align-items: center;
-  
-    div {
-      display: flex;
-      gap: 1rem;
-  
-      a {
-        color: $dark;
-      }
-    }
-  
-    h1 {
-      font-size: 1.5rem;
-      color: $dark;
-      font-weight: bold;
+    margin-bottom: 1rem;
+    gap: 0.5rem;
+}
+
+.storage-header-link {
+    background: none;
+    border: none;
+    outline: none;
+}
+
+.storage-header-link[data-selected] {
+    color: $primary;
+
+    &:hover {
+        color: $primary-hover;
     }
 }
